+++ .env
@@ -1,0 +1,3 @@
+PORT=3001
+URL=mongodb://localhost:27017/person
+SECRET="anystring"
