--- src/.env
@@ -1,1 +1,0 @@
-PORT=3000
