+++ src/index.ts
@@ -1,8 +1,8 @@
 import express from "express";
 import mongoose from "mongoose";
-import dbConnect from "./config/connection";
-import userdetail from "./model/userdetail";
-import router from "./router/router";
+import dbConnect from "./config/user.connection";
+import userdetail from "./model/user.detail";
+import router from "./user.router/user.router";
 import * as dotenv from "dotenv";
 
 dotenv.config();
@@ -13,7 +13,7 @@ dbConnect;
 userdetail;
 app.use(router);
 
-app.listen(process.env.PORT||3000, () => {
+app.listen(process.env.PORT, () => {
   console.log(process.env.PORT);
   console.log(`Express server is running.`);
 });
