+++ src/middleware/user.auth.ts
@@ -1,19 +1,20 @@
 import express, { application } from "express";
-import userdetail from "../model/userdetail";
+import userdetail from "../model/user.detail";
 import bycrypt from "bcryptjs";
 import jwt from "jsonwebtoken";
 import { create } from "ts-node";
 import { Request, Response, NextFunction } from "express";
+const key=String(process.env.SECRET);
 
 const register = async(req: Request, res: Response)=> {
   const { username, password, firstname, lastname, gmail, phoneNumber } =
     req.body;
   const userExist = await userdetail.findOne({ username });
   if (userExist) {
-    res.json({ error: "User already Exist" });
+    res.json({ error: "User already exist." });
   } else {
-    const salt = await bycrypt.genSalt(10);
-    const hashpassword = await bycrypt.hash(password, salt);
+    const salt:string = await bycrypt.genSalt(10);
+    const hashpassword:string = await bycrypt.hash(password, salt);
     const userdetail1 = new userdetail({
       username: username,
       password: hashpassword,
@@ -30,7 +31,7 @@ const register = async(req: Request, res: Response)=> {
       };
       jwt.sign(
         payload,
-        "anystring",
+        key,
         { expiresIn: 2 * 24 * 60 * 1000 },
         function (err, token) {
           if (err) {
@@ -57,11 +58,10 @@ const login= async (req:Request, res:Response) => {
         const payload = {
           
             _id: data._id,
-         // },
         };
         jwt.sign(
           payload,
-          "anystring",
+          key,
           { expiresIn: 2 * 60 * 10000 },
           function (err, token) {
             if (err) {
@@ -92,15 +92,17 @@ const login= async (req:Request, res:Response) => {
     } catch (err) {
       return res.status(401).send("Invalid Token");
     }
+    res.json({message:"Welcome! Token verified."})
     return next();
+  
 };
 
  const getProfile=async (req:Request, res:Response) => {
   if (req.headers && req.headers.authorization) {
     const authorization = req.headers.authorization;
     try {
-      const decoded= jwt.verify(authorization, "anystring") as { _id : string};
-      const userId= decoded._id;
+      const decoded= jwt.verify(authorization, key) as { _id : string};
+      const userId:string= decoded._id;
       const detail = await userdetail.findOne({ _id: userId }).lean();
       //lean retruns json object.
       res.json(detail);
@@ -114,9 +116,9 @@ const updateProfile= async (req:Request, res:Response) => {
   if (req.headers && req.headers.authorization) {
     var authorization = req.headers.authorization;
     try {
-      const decoded = jwt.verify(authorization, "anystring") as { _id : string};
+      const decoded = jwt.verify(authorization, key) as { _id : string};
       console.log(decoded);
-      var userId = decoded._id;
+      const userId:string = decoded._id;
       const detail = await userdetail.updateOne(
         { _id: userId },
         { $set: req.body }
@@ -133,15 +135,15 @@ const updateProfile= async (req:Request, res:Response) => {
 const deactivate= async (req:Request, res:Response) => {
   if (req.headers && req.headers.authorization) {
     var authorization = req.headers.authorization;
-    const decoded: any = jwt.verify(authorization, "anystring")as { _id : string};
-    var userId = decoded._id;
+    const decoded: any = jwt.verify(authorization, key)as { _id : string};
+    var userId:string = decoded._id;
     console.log(userId);
     const data = await userdetail.findOneAndUpdate(
       { _id: userId },
       { $set: { is_active: false } }
     );
     res.json(data);
-  } else res.json({message:"token not matched."});
+  } else res.json({message:"Token not matched."});
 };
 
 const reactiveProfile=async (req:Request, res:Response) => {
@@ -150,11 +152,9 @@ const reactiveProfile=async (req:Request, res:Response) => {
   let data = await userdetail.findOne({ username });
   if (data && bycrypt.compare(password, data.password)) {
     const payload = {
-      userdetail: {
-        _id: data?._id,
-      },
+        _id: data?._id
     };
-    const token = jwt.sign(payload, "anystring", {
+    const token:string = jwt.sign(payload, key, {
       expiresIn: 2 * 60 * 10000,
     });
     const Data = await userdetail.updateOne(
