+++ src/model/user.interface.ts
@@ -1,0 +1,11 @@
+interface userdetail {
+    username: String;
+    password: string;
+    firstname: String;
+    lastname: String;
+    gmail: String;
+    phoneNumber: Number;
+    is_active: Boolean;
+  }
+
+  export default userdetail;
